Dashboard working at 95%, need to solve issue with how to load history data, might need to refactor backend to send specific data instead of the bulk, that might make it faster

+++ pages/api/addDish.js
@@ -13,7 +13,6 @@ export const config = {
 
 export default async (req, res) => {
   const data = await new Promise((resolve, reject) => {
-    // console.log(path.join(`${__dirname}/../../../../`, 'public/'))
     const form = new formidable({
       uploadDir: path.join(`${__dirname}/../../../../`, 'public'),
       keepExtensions: true,
@@ -24,8 +23,11 @@ export default async (req, res) => {
       resolve({ err, fields, files });
     });
   });
-
-  let fileName = `dish-${data.files.image.name}-${Date.now()}.jpeg`;
+  let fileName = 'stockDishImg.png';
+  if (Object.keys(data.files).length > 0) {
+    fileName = `dish-${data.files.image.name}-${Date.now()}.jpeg`;
+  }
+  // console.log(data);
 
   const addDish = await fetch('http://localhost:3001/api/v1/menu', {
     method: 'POST',
@@ -46,30 +48,35 @@ export default async (req, res) => {
   const result = await addDish.json();
 
   if (addDish.ok) {
-    var oldpath = data.files.image.path;
-    var modifiedPath = `${oldpath.split('.')[0]}-new.${oldpath.split('.')[1]}`;
-    sharp(oldpath)
-      .resize(500, 500)
-      .toFormat('jpeg')
-      .jpeg({ quality: 90 })
-      .toFile(modifiedPath)
-      .then(() => {
-        var newpath =
-          path.join(`${__dirname}/../../../../`, 'public') +
-          '/dishes/' +
-          fileName;
+    if (Object.keys(data.files).length > 0) {
+      var oldpath = data.files.image.path;
+      var modifiedPath = `${oldpath.split('.')[0]}-new.${
+        oldpath.split('.')[1]
+      }`;
+      sharp(oldpath)
+        .resize(500, 500)
+        .toFormat('jpeg')
+        .jpeg({ quality: 90 })
+        .toFile(modifiedPath)
+        .then(() => {
+          var newpath =
+            path.join(`${__dirname}/../../../../`, 'public') +
+            '/dishes/' +
+            fileName;
 
-        fs.rename(modifiedPath, newpath, function (err) {
-          if (err) throw err;
+          fs.rename(modifiedPath, newpath, function (err) {
+            if (err) throw err;
+          });
+        })
+        .then(() => {
+          fs.unlink(data.files.image.path, (err) => {
+            if (err) {
+              throw err;
+            }
+            // console.log('File deleted')
+          });
         });
-      }).then(() => {
-        fs.unlink(data.files.image.path, (err) => {
-          if (err) {
-            throw err;
-          }
-          // console.log('File deleted')
-        });
-      });
+    }
 
     res.status(201).json({
       status: 'success',
@@ -78,12 +85,13 @@ export default async (req, res) => {
       },
     });
   } else {
+    if(Object.keys(data.files).length > 0){
     fs.unlink(data.files.image.path, (err) => {
       if (err) {
         throw err;
       }
       // console.log('File deleted')
-    });
+    });}
     res.status(401).json({
       status: 'failed',
       data: {
